pyproject.toml
[project]
name = "intelehealth-ai-models"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "dspy==2.5.39",
    "google-generativeai==0.8.3",
    "python-dotenv>=0.4.0",
    "uvicorn>=0.34.0",
    "requests>=2.32.3",
    "mlflow>=3.1.1",
    "pymysql>=1.1.1",
    "fastapi>=0.111.0",
    "pydantic>=2.7.1",
    "groq>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modules", "signatures", "utils", "ops"]
